refactor: lib overhaul
so this bases lib off of nixpkgs lib, the most important part to note
here is significant API changes, perticuallarly when it comes to
builders. in order to slay some IR demons we needed to change how the
builders worked since using `flake.x` in the way I was is *not* amazing,
but this fixes some consistancy issues but makes darwin vs nixos
configurations look a bit uglier then before.